waitfor err: kill subprocs and exit immediately
The possible errors returned by wait(2) are ECHILD (no child processes)
and EINTR.

In the case of EINTR, the wait(2) is retried.

The only situation in which there would be no child processes is if
fork(2) had previously failed (and the program would have exited).

Crash if wait(2) errors out and terminate any subprocs. Don't write the
exit status to the lockfile.