Run ccronexpr in a subprocess
cron expressions wouldn't normally come from an untrusted source. But an
invalid expression can cause stack overflows (for example, cron_prev()
with local time during a time change), get stuck in an infinite loop or
have buffer overflows.

Because runcron needs to fork()/exec() the cron job, the runcron
process itself can't be restricted (the subprocess would inherit the
restrictions).

To isolate ccronexpr, fork() and run it in a subprocess. Process
restrictions can be applied to the subprocess only.

Note: OpenBSD pledge(2) is the exception. With execpromises, the fork()
can possibly be avoided.