rlimit: restrict process using setrlimit(2)
When rlimit mode process restriction is enabled, the cron parsing
subprocess is not allowed to fork(), open new files or write to any open
file descriptors that write to the filesystem.

Further setrlimit() restrictions could be applied to all modes (even the
null sandbox) to limit memory and cpu usage so a buggy cron expression
does hang in an infinite loop.