Restrict the runtime of the cron parsing
To avoid situations where a cron expression causes an infinite loop or
allocates a large amount of memory, restrict the cron expression parsing
subprocess to an arbitrary 10 seconds of CPU and 1Mb of memory.

These values were arbitrarily chosen after light testing. They could be
user configurable but really just need to be an upper bound on
runtime/memory usage.