runcron: account for SIGKILL
During these conditions:

* the previous runcron job was successful
* runcron has forked the task and is blocking for the exit status
* runcron is sent an uncatchable signal

The next invocation of runcron will mark the task exit as a success and
wait for the next cron interval since the state file was not updated:

    runcron -vvv '*/10 * * * * *' echo test
    # exit status=0
    runcron -vvv '*/10 * * * * *' sleep inf
    # kill -9 <runcronpid>
    # exit status=0

This situation might occur, for example, during container shutdown or
during VM or hardware shutdown. Note receipt of SIGKILL may leave
task subprocesses running.

If the last exit status was a success, update the state file to indicate
SIGKILL before running the task. The state will be overwritten with the
actual exit value when runcron exits.

In the case where the last exit status is non-0, the state file is not
changed: the task will run anyway on next invocation.