openbsd: adjust --timestamp seconds for timezone
The --timestamp option is used for testing. Some versions of strptime(3)
accept the "%s" format option or number of seconds since the Epoch.

The OpenBSD strptime(3) "%s" format option returns a time value in UTC
without accounting for the local time zone. Fix by passing in the local
time value and adding the time zone offset.