aliases: @monthly: limit to 28 days
The @monthly alias randomly selected a time between the 1st and 31st of
the month. Months with less than 31 days could be skipped if the
selected day exceeded the number of days in the month.

To force aliases to choose a predictable time, fix @monthly by restricting
the choice to the first 28 days.

Using a variable number of days for the month could result in a job that
runs on the 30th or 31st day.