capsicum: use process descriptors
Drop capabilities while waiting for the subprocess. As a consequence,
the parent process can only signal processes to which it holds a file
descriptor which breaks terminating background processes.

TODO: investigate if a process in capability mode can use procctl(2) to
set itself as a reaper and terminate all subprocesses on exit.